# team-17c-main/user-microservice/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: User Microservice API
  description: >
    The User Microservice allows users to perform different operations on their accounts, while specifying 
    their own user_id in a header named 'user_id' where needed. The root address of this API is '/user'.
    
    
    Certain operations are only executable by users which hold a certain role, for instance, the '/admin' paths are
    only executable by users which obtained the 'Admin' role. 
    The role hierarchy is (Regular) User < Author < Admin when comparing 
    by permissions. Regular users and authors are not able to view other users' credentials, in fact, they are 
    only allowed to view other users' profiles (which do not contain sensitive attributes). Nobody is allowed to view 
    their own nor other users' passwords through the endpoints, modification of one's password is possible.
    
    
    Admins are users too in our application.

  version: 1.0.0
servers:
  - url: http://localhost:8080/user
...
components:
  parameters:
    UserIDHeader: # try using this when a header is specified with the user email
      name: user_id
      in: header
      description: Your User ID
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    UserProfile:
      required:
        - username
      type: object
      x-class-extra-annotation : "@javax.persistence.Entity"
      properties:
        # needed for usage as follower/following and finding user profile by user id (to prevent
        # other users from seeing your email/password)
        user_id:
          description: the ID of the user
          type: string
          format: uuid
          x-field-extra-annotation: "@javax.persistence.Id @javax.persistence.GeneratedValue"

#          x-annotations:
#            - "@javax.persistence.Id"
#            - "@GeneratedValue(generator = \"uuid2\")"
#            - "@GenericGenerator(name = \"uuid2\", strategy = \"org.hibernate.id.UUIDGenerator\")"
#            - "@Column(name = \"userId\", columnDefinition = \"VARCHAR(255)\")"
        username:
          type: string
          pattern: "^\\w{2,20}$"
#        email:
#          type: string
#          format: email
        name:
          type: string
        privacy:
          type: integer
          enum:
            - 0
            - 1
          x-enum-varnames:
            - PUBLIC
            - PRIVATE
          description: >
            Privacy ID:
             * `0` - Public Profile
             * `1` - Private Profile
        role:
          type: integer
          enum:
            - 0
            - 1
            - 2
          x-enum-varnames:
            - USER
            - AUTHOR
            - ADMIN
          description: >
            Role ID:
             * `0` - User Role
             * `1` - Author Role
             * `2` - Admin Role
        state:
          type: integer
          enum:
            - 0
            - 1
            - 2
          x-enum-varnames:
            - ACTIVE
            - INACTIVE
            - BANNED
          default: 0  # Set the default value to ACTIVE
          description: >
            State ID:
             * `0` - Active
             * `1` - Inactive
             * `2` - Banned
        bio:
          type: string
        avatar:
          description: The profile picture
          type: string
        location:
          type: string
        favoriteGenres:
          description: array of ids of favorite genres
          type: array
          x-field-extra-annotation: "@javax.persistence.ElementCollection"
          items:
            type: string
        favoriteBook:
          description: id of favorite book
          type: string

        # for followers and following, what if 2 users follow each other
        # let's say user1 and user2, what if we look at the profile of user1, this will
        # show user2 in the following/followers list, then user2 will show user1 in the following/followers list
        # and so on... doesn't this implementation create infinite loops?
        # maybe we can just store the user id's, but that's just problems for the code itself
        # we can leave this object as it is right now, since it doesn't matter for the OpenAPI stuff
        followers:
          type: array
          description: UserProfiles of followers
          x-field-extra-annotation: "@javax.persistence.ElementCollection"
          items:
            description: the ID of the profile
            type: string
            format: uuid
        following:
          type: array
          description: UserProfiles of the "followed"
          x-field-extra-annotation: "@javax.persistence.ElementCollection"
          items:
            description: the ID of the profile
            type: string
            format: uuid
        friends:
          type: array
          description: UserProfiles of friends, friends = users who follow each other back
          x-field-extra-annotation: "@javax.persistence.ElementCollection"
          items:
            description: the ID of the profile
            type: string
            format: uuid
        bookshelves:
          description: array of id's of the bookshelves
          type: array
          x-field-extra-annotation: "@javax.persistence.ElementCollection"
          items:
            type: string
            format: UUID
...
paths:
...
  /profile/view:
    get:
      tags:
        - UserProfile
      summary: Get your own profile.
      description: Get the UserProfile of current user.
      operationId: getOwnProfile
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized, invalid/missing ID in Header
  /profile/modify:
    put:
      tags:
        - UserProfile
      summary: Update your own profile.
      description: Update the UserProfile of the current user.
      operationId: updateOwnProfile
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Bad Request
        '401':
          description: Unauthorized, invalid/missing ID in Header
        '405':
          description: Invalid Input
  /profile/view/{user_id}:
    get:
      tags:
        - UserProfile
      summary: View the profile of a user.
      description: View the UserProfile of a user specified by user_id, if viewable.
      operationId: viewProfile
      parameters:
        - name: user_id
          in: path
          description: User ID of profile to view
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid ID supplied
        '404':
          description: User Profile not found
        '401':
          description: Unauthorized operation
  /profile/all:
    get:
      tags:
        - UserProfile
      summary: View all user profiles.
      description: View all (viewable) UserProfile's in the system.
      operationId: viewAllProfiles
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfile'
          description: Successful operation
        '400':
          description: Bad Request
  /profile/find:
    post:
      tags:
        - UserProfile
      summary: Get all user profiles that meet criteria.
      description: Get all the user profiles that meet the criteria in the request body, like name, interests etc.
      operationId: findProfiles
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfile'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserProfile'
        '400':
          description: Bad request
        '404':
          description: Users not found
  /profile/deactivate:
    put:
      tags:
        - UserProfile
      summary: Deactivate your account.
      description: Toggle your account status to inactive.
      operationId: deactivateAccount
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
      responses:
        '200':
          description: Successfully Deactivated
        '400':
          description: Bad Request
        '405':
          description: Invalid input
  /profile/activate:
    put:
      tags:
        - UserProfile
      summary: Activate your account.
      description: Toggle your account status to active.
      operationId: activateAccount
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
      responses:
        '200':
          description: Successfully Activated
        '400':
          description: Bad Request
        '401':
          description: User is prohibited
        '405':
          description: Invalid input
  /profile/proveAuthor:
    post:
      tags:
        - UserProfile
      summary: Prove you're an author with an author keyphrase.
      description: Provide a string which contains the word "author" (case insensitive) to prove author identity.
      operationId: proveAuthor
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: string
              example: I am an author

      responses:
        '200':
          description: Successfully Activated
        '400':
          description: Bad Request
        '401':
          description: Author credentials invalid or User prohibited
        '405':
          description: Invalid input
  /profile/proveAdmin:
    post:
      tags:
        - UserProfile
      summary: Prove you're an admin with an admin key.
      description: Prove you're an admin using an admin key, which has been generated by another admin.
      operationId: proveAdmin
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
      requestBody:
        content:
          application/json:
            schema:
              type: string
              format: uuid
      responses:
        '200':
          description: Successfully Proved Admin
        '400':
          description: Bad Request
        '401':
          description: Admin Password Invalid or User prohibited
        '405':
          description: Invalid Input
...
  /profile/follow/{user_id_other}:
    put:
      operationId: followUser
      tags:
        - UserProfile
      summary: Follow a user.
      description: Follow another user, specified by user_id.
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
        - name: user_id_other
          in: path
          description: User ID of user to follow
          required: true
          schema:
            type: string
            format: uuid

      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
        '401':
          description: Unauthorized operation
  /profile/unfollow/{user_id_other}:
    delete:
      operationId: unfollowUser
      tags:
        - UserProfile
      summary: Unfollow a user.
      description: Unfollow another user, specified by user_id.
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
        - name: user_id_other
          in: path
          description: User ID of user to unfollow
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid ID supplied
        '404':
          description: User not found
        '401':
          description: Unauthorized operation
  /profile/favoriteBook/{book_id}:
    put:
      operationId: favoriteBook
      tags:
        - UserProfile
      summary: Add book to favorites.
      description: Add your favorite book, specified by book_id, as a user, possibly replacing the previous entry.
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
        - name: book_id
          in: path
          description: Book ID of book selected as favorite.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Book not found
        '401':
          description: Unauthorized operation
  /profile/unfavoriteBook/{book_id}:
    delete:
      operationId: unfavoriteBook
      tags:
        - UserProfile
      summary: Delete book from favourites.
      description: Delete your favorite book as a user, specified by book_id.
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
        - name: book_id
          in: path
          description: Book ID of book selected to delete as a favorite.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Book not found
        '401':
          description: Unauthorized operation
  /profile/favoriteGenre/{genre_id}:
    put:
      operationId: favoriteGenre
      tags:
        - UserProfile
      summary: Add genre to favourites.
      description: Add your favorite genre as a user, specified by genre_id.
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
        - name: genre_id
          in: path
          description: Genre ID of genre selected as favorite.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Genre not found
        '401':
          description: Unauthorized operation
  /profile/unfavoriteGenre/{genre_id}:
    delete:
      operationId: unfavoriteGenre
      tags:
        - UserProfile
      summary: Delete genre from favourites.
      description: Delete one of your favorite genres as a user, specified by genre_id.
      parameters:
        - $ref: '#/components/parameters/UserIDHeader'
        - name: genre_id
          in: path
          description: Genre ID of genre selected to delete as a favorite.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid ID supplied
        '404':
          description: Genre not found
        '401':
          description: Unauthorized operation